Add random identifier and changed master key structure
1 - To allow a safe deletion of blobs, a random identifier is associated to each of the. It is saved in the backend with the blob and it is required to delete the blob.

2 - MasterKey used to be the result of encrypting a record containing the references for index and user preferences with the master password. The master password is vulnerable and encrypting a content that has the same structure because of its json representation could exploited.
A new structure called UserInfo is introduced, containing the index reference and the user preferences. To compute the new MasterKey, the UserInfo in encrypted with a new random key. The hash of the encrypted UserInfo is concatenated with the key itself and the result is then encrypted with the master password.