Removed development-warning
The warning has been moved to an external page that will embed the application in an iFrame. In this way we can adopt the same solution both for the "preview" warning for the new app, and the "legacy" warning for the old version once the new version will be online.